feat: 홈 v2 메인탭 연결
