+++ Projects/Features/HomeFeatureV2/Demo/Sources/HomeFeatureDemoApp.swift
@@ -1,25 +1,37 @@
 import ComposableArchitecture
+import Data
 import Domain
 import HomeFeatureV2
 import SwiftUI
 
 @main
 struct HomeFeatureV2DemoApp: App {
+    
+    private let demoUserUuid = Bundle.main.object(
+        forInfoDictionaryKey: "HOME_DEMO_USER_UUID"
+    ) as? String ?? ""
+    private let homePopupClient = HomePopupClient.live(
+        popupUsecase: PopupUsecaseImpl(
+            popupRepository: PopupRepositoryImpl()
+        )
+    )
+    
+    init() {
+        print("테스트: \(demoUserUuid)")
+    }
+    
     var body: some Scene {
         WindowGroup {
-
-            // let popups: [Popup] = Array(repeating: .popupMock, count: 20)
-            
             HomeFeatureView(
                 store: Store(
                     initialState: HomeFeature.State(
                         user: User(
-                            userUuid: "preview-user",
+                            userUuid: demoUserUuid,
                             uid: "preview-uid",
                             provider: "preview",
                             email: nil,
                             nickname: "홍길동",
-                            role: "USER",
+                            role: "ADMIN",
                             isAlerted: false,
                             fcmToken: nil,
                             alertKeywordList: nil,
@@ -28,11 +40,10 @@ struct HomeFeatureV2DemoApp: App {
                     )
                 ) {
                     HomeFeature()
-                } withDependencies: {
-                    $0.homePopupClient = .previewValue
-//                    $0.homePopupClient.getPersonalRandomPopupList = { _ in
-//                        Array(repeating: .popupMock, count: 20)
-//                    }
+                }
+                withDependencies: {
+                    $0.homePopupClient = homePopupClient
+                    // $0.homePopupClient = .previewValue
                 }
             )
         }
@@ -43,6 +54,10 @@ struct HomeFeatureV2DemoApp: App {
 
 
 
+
+//                    $0.homePopupClient.getPersonalRandomPopupList = { _ in
+//                        Array(repeating: .popupMock, count: 20)
+//                    }
 
 //#Preview {
 //    HomeFeatureView(
